[chore] #60 공통 컴포넌트화
