+++ app/build.gradle.kts
@@ -110,8 +110,7 @@ dependencies {
     implementation(libs.bundles.kakao)
 
     // View Pager
-    implementation(libs.accompanist.pager)
-    implementation(libs.accompanist.pager.indicators)
+    implementation(libs.bundles.pager)
 }
 
 ktlint {
