+++ app/src/main/java/org/sopt/dateroad/presentation/ui/onboarding/OnboardingScreen.kt
@@ -2,18 +2,13 @@ package org.sopt.dateroad.presentation.ui.onboarding
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
@@ -22,11 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
@@ -36,6 +27,8 @@ import kotlinx.coroutines.launch
 import org.sopt.dateroad.R
 import org.sopt.dateroad.presentation.type.OnboardingType
 import org.sopt.dateroad.presentation.ui.component.button.DateRoadFilledButton
+import org.sopt.dateroad.presentation.ui.component.partialcolortext.PartialColorText
+import org.sopt.dateroad.presentation.ui.onboarding.component.DotsIndicator
 import org.sopt.dateroad.ui.theme.DateRoadTheme
 
 @OptIn(ExperimentalPagerApi::class)
@@ -75,7 +68,11 @@ fun OnboardingScreen(
                 )
                 Spacer(modifier = Modifier.height(22.dp))
                 Text(
-                    text = getStyledText(stringResource(id = onboardingType.titleRes)),
+                    text = PartialColorText(
+                        stringResource(id = onboardingType.titleRes),
+                        keywords = listOf("포인트", "데이트코스", "100", "다양한"),
+                        color = DateRoadTheme.colors.deepPurple
+                    ),
                     style = DateRoadTheme.typography.titleExtra24,
                     textAlign = TextAlign.Center
                 )
@@ -97,9 +94,7 @@ fun OnboardingScreen(
         }
         Spacer(modifier = Modifier.height(14.dp))
         DateRoadFilledButton(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 60.dp),
+            modifier = Modifier.align(Alignment.CenterHorizontally).width(242.dp),
             isEnabled = true,
             textContent =
             if (pagerState.currentPage == OnboardingType.entries.size - 1) {
@@ -130,45 +125,10 @@ fun OnboardingScreen(
         DotsIndicator(
             totalDots = OnboardingType.entries.size,
             selectedIndex = pagerState.currentPage,
-            modifier = Modifier.align(Alignment.CenterHorizontally)
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+                .padding(horizontal = 158.dp)
         )
         Spacer(modifier = Modifier.height(15.dp))
-    }
-}
-
-@Composable
-fun DotsIndicator(totalDots: Int, selectedIndex: Int, modifier: Modifier = Modifier) {
-    Row(
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.Bottom,
-        modifier = modifier.width(44.dp)
-    ) {
-        for (i in 0 until totalDots) {
-            val color = if (i == selectedIndex) DateRoadTheme.colors.deepPurple else DateRoadTheme.colors.gray200
-            Box(
-                modifier = Modifier
-                    .size(8.dp)
-                    .background(color, CircleShape)
-            )
-        }
-    }
-}
-
-@Composable
-fun getStyledText(text: String): AnnotatedString {
-    val keywords = listOf("포인트", "데이트코스", "100", "다양한")
-    return buildAnnotatedString {
-        var currentIndex = 0
-        keywords.forEach { keyword ->
-            val keywordIndex = text.indexOf(keyword, currentIndex, ignoreCase = true)
-            if (keywordIndex >= 0) {
-                append(text.substring(currentIndex, keywordIndex))
-                withStyle(style = SpanStyle(color = DateRoadTheme.colors.deepPurple)) {
-                    append(text.substring(keywordIndex, keywordIndex + keyword.length))
-                }
-                currentIndex = keywordIndex + keyword.length
-            }
-        }
-        append(text.substring(currentIndex))
     }
 }
